STS.h: Declare a couple of vars as int.
They are used as such already.

+++ src/Subtitles/STS.h
@@ -44,9 +44,9 @@ public:
     double   fontScaleX, fontScaleY; // percent
     double   fontSpacing;    // +/- pixels
     LONG     fontWeight;
-    BYTE     fItalic;
-    BYTE     fUnderline;
-    BYTE     fStrikeOut;
+    int      fItalic;
+    int      fUnderline;
+    int      fStrikeOut;
     int      fBlur;
     double   fGaussianBlur;
     double   fontAngleZ, fontAngleX, fontAngleY;
